requirements.txt
pdfminer.six>=20240706    # primary PDF text extractor for the pre-OCR cache
pytesseract>=0.3.10       # optional — OCR fallback for scanned PDFs (needs system tesseract)
pdf2image>=1.17           # optional — only needed alongside pytesseract
pypdf>=4.0                # optional — backup PDF text extractor
